+++ decidim-elections/spec/system/admin_manages_answers_spec.rb
@@ -3,6 +3,8 @@
 require "spec_helper"
 
 describe "Admin manages answers", type: :system do
+  let!(:proposals) { create_list :proposal, 3, :accepted, component: origin_component }
+  let!(:origin_component) { create :proposal_component, participatory_space: current_component.participatory_space }
   let(:election) { create :election, component: current_component }
   let(:question) { create :question, election: election }
   let(:answer) { create :election_answer, question: question }
@@ -22,6 +24,21 @@ describe "Admin manages answers", type: :system do
 
     within find("tr", text: translated(question.title)) do
       page.find(".action-icon--edit-answers").click
+    end
+  end
+
+  describe "importing proposals" do
+    it "imports proposals" do
+      click_on "Import proposals to answers"
+
+      within ".import_proposals" do
+        select origin_component.name["en"], from: :proposals_import_origin_component_id
+        check :proposals_import_import_all_accepted_proposals
+      end
+
+      click_button "Import proposals to answers"
+
+      expect(page).to have_content("3 proposals successfully imported")
     end
   end